Refactor BottomNavView to use Flutter Hooks and add ShopView
This commit refactors the BottomNavView to use Flutter Hooks instead of being a StatelessWidget. It also adds the ShopView to the list of screens that can be navigated to using the BottomNavigationBar. The selected screen is now stored in a ValueNotifier and updated when the user taps on a different tab.